Buffer records when execting query without block.
