+++ lib/db/records.rb
@@ -1,0 +1,58 @@
+# frozen_string_literal: true
+
+# Copyright, 2020, by Samuel G. D. Williams. <http://www.codeotaku.com>
+# 
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+# 
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+# 
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+module DB
+	# A buffer of records.
+	class Records
+		EMPTY = self.new.freeze
+		
+		def self.wrap(result)
+			# We want to avoid extra memory allocations when there are no columns:
+			if result.field_count == 0
+				return EMPTY
+			end
+			
+			return self.new(result.field_names, result.to_a)
+		end
+		
+		def initialize(columns, rows)
+			@columns = columns
+			@rows = rows
+		end
+		
+		def freeze
+			return self if frozen?
+			
+			@columns.freeze
+			@rows.freeze
+			
+			super
+		end
+		
+		attr :columns
+		attr :rows
+		
+		def to_a
+			@rows
+		end
+	end
+end
